% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafiles.R
\docType{data}
\name{abdata}
\alias{abdata}
\title{Employment UK data}
\format{The variables are:
\describe{
  \item{c1}{Record ID}
  \item{ind}{Firm index}
  \item{year}{Year}
  \item{emp}{Employment}
  \item{wage}{Wage}
  \item{cap}{Capital}
  \item{indoutpt}{Industrial output}
  \item{n, w, k, ys}{Logs of variables}
  \item{rec}{Record number}
  \item{yearm1}{Lagged year}
  \item{id}{ID}
  \item{nL1, nL2, wL1, kL1, kL2, ysL1, ysL2}{Lags of log variables}
  \item{yr1976 - yr1984}{Time dummies}
}}
\source{
\url{http://www.stata-press.com/data/r13/abdata.dta}
}
\usage{
abdata
}
\description{
This data set contains labor demand data from a panel of firms in the United Kingdom. The panel is unlanced.
}
\references{
Arellano, M. and Bond, S. (1991) "Some tests of specification for panel data: Monte Carlo evidence and an application to employment equations", \emph{The Review of Economic Studies}, \strong{58}(2), 227-297, \href{http://dx.doi.org/10.2307/2297968}{doi: 10.2307/2297968}
}
\keyword{datasets}
