% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{Andrews_Lu_MMSC}
\alias{Andrews_Lu_MMSC}
\alias{Andrews_Lu_MMSC.pvargmm}
\title{Andrews Lu MMSC Criteria based on Hansen-J-Statistic}
\usage{
Andrews_Lu_MMSC(model, HQ_criterion = 2.1)

\method{Andrews_Lu_MMSC}{pvargmm}(model, HQ_criterion = 2.1)
}
\arguments{
\item{model}{A PVAR model}

\item{HQ_criterion}{Hannan Quinn criterion}
}
\value{
BIC, AIC and HQIC
}
\description{
...
}
\examples{
data("ex3_abdata")
Andrews_Lu_MMSC(ex3_abdata)



}
\references{
Andrews, D., Lu, B. (2001) Consistent Model and Momement Selection Procedures for GMM Estimation with Application to Dynamic Panel Data Models, \emph{Journal of Econometrics}, \bold{101}(1), 123--164, \doi{10.1016/S0304-4076(00)00077-4}
}
