% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvarfeols-methods.R
\name{coef.pvarfeols}
\alias{coef.pvarfeols}
\title{Extract PVARFEOLS(p) Model Coefficients}
\usage{
\method{coef}{pvarfeols}(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{further arguments}
}
\description{
Extract PVARFEOLS(p) Model Coefficients
}
