% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvar-methods.R
\name{coef.pvargmm}
\alias{coef.pvargmm}
\title{Extract PVAR(p) Model Coefficients}
\usage{
\method{coef}{pvargmm}(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{further arguments}
}
\description{
Extract PVAR(p) Model Coefficients
}
\examples{
data("ex1_dahlberg_data")
coef(ex1_dahlberg_data)
}
