% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvarhk-methods.R
\name{coef.pvarhk}
\alias{coef.pvarhk}
\title{Extract PVARHK(p) Model Coefficients}
\usage{
\method{coef}{pvarhk}(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{further arguments}
}
\description{
Extract PVARHK(p) Model Coefficients
}
