% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafiles.R
\docType{data}
\name{Dahlberg}
\alias{Dahlberg}
\title{Swedish municipalities data}
\format{The variables are:
\describe{
  \item{id}{ID number for municipality}
  \item{year}{Year}
  \item{expenditures}{Total expenditures}
  \item{revenues}{Total own-source revenues}
  \item{grants}{Intergovernmental grants received by the municipality}
}

Total expenditures contains both capital and current expenditures.

Expenditures, revenues, and grants are expressed in million SEK. The
series are deflated and in per capita form. The implicit deflator is a
municipality-specific price index obtained by dividing total local
consumption expenditures at current prices by total local consumption
expenditures at fixed (1985) prices.

The data are gathered by Statistics Sweden and obtained from
Financial Accounts for the Municipalities (Kommunernas Finanser).}
\source{
\url{http://qed.econ.queensu.ca/jae/2000-v15.4/dahlberg-johansson/}
}
\usage{
Dahlberg
}
\description{
The panel data set consists of 265 Swedish municipalities and
covers 9 years (1979-1987).
}
\references{
M. Dahlberg and E. Johansson (2000) "An examination of the dynamic behavior of local governments using GMM bootstrapping methods", \emph{Journal of Applied Econometrics}, \strong{15}(4), 401-416, \url{http://www.jstor.org/stable/2678589}.
}
\keyword{datasets}
