% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafiles.R
\docType{data}
\name{ex1_dahlberg_data}
\alias{ex1_dahlberg_data}
\title{Dahlberg results example 1}
\format{An object of class \code{pvargmm} of length 34.}
\usage{
ex1_dahlberg_data
}
\description{
Dahlberg results example 1
}
\keyword{datasets}
