% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafiles.R
\docType{data}
\name{ex1_dahlberg_data_bs}
\alias{ex1_dahlberg_data_bs}
\title{Dahlberg bootstrap results example 1}
\format{An object of class \code{list} of length 4.}
\usage{
ex1_dahlberg_data_bs
}
\description{
Dahlberg bootstrap results example 1
}
\keyword{datasets}
