% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafiles.R
\docType{data}
\name{ex2_nlswork2_data_bs}
\alias{ex2_nlswork2_data_bs}
\title{NLS Work 2 bootstrap results example 2}
\format{
An object of class \code{list} of length 4.
}
\usage{
ex2_nlswork2_data_bs
}
\description{
NLS Work 2 bootstrap results example 2
}
\keyword{datasets}
