% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafiles.R
\docType{data}
\name{ex3_abdata}
\alias{ex3_abdata}
\title{Example results for Employment UK data}
\format{An object of class \code{pvargmm} of length 34.}
\usage{
ex3_abdata
}
\description{
Example results for Employment UK data
}
\keyword{datasets}
