% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvar-methods.R, R/pvarfeols-methods.R,
%   R/pvarhk-methods.R
\name{extract}
\alias{extract}
\alias{extract.pvargmm}
\alias{extract.pvarfeols}
\alias{extract.pvarhk}
\title{Extract Coefficients and GOF Measures from a Statistical Object}
\usage{
extract(model, ...)

\method{extract}{pvargmm}(model, ...)

\method{extract}{pvarfeols}(model, ...)

\method{extract}{pvarhk}(model, ...)
}
\arguments{
\item{model}{Model}

\item{...}{Further arguments passed to or from other methods}
}
\description{
Extract Coefficients and GOF Measures from a Statistical Object
}
\examples{
data("ex1_dahlberg_data")
extract(ex1_dahlberg_data)
}
