% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvar-methods.R
\name{fixedeffects}
\alias{fixedeffects}
\alias{fixedeffects.pvargmm}
\title{Extracting Fixed Effects}
\usage{
fixedeffects(model, ...)

\method{fixedeffects}{pvargmm}(model, Only_Non_NA_rows = TRUE, ...)
}
\arguments{
\item{model}{Model}

\item{...}{Further arguments passed to or from other methods}

\item{Only_Non_NA_rows}{Filter NA rows}
}
\description{
Extracting Fixed Effects
}
\examples{
data("ex1_dahlberg_data")
fixedeffects(ex1_dahlberg_data)
}
