% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvar-methods.R
\name{hansen_j_test}
\alias{hansen_j_test}
\alias{hansen_j_test.pvargmm}
\title{Sargan-Hansen-J-Test for Overidentification}
\usage{
hansen_j_test(model, ...)

\method{hansen_j_test}{pvargmm}(model, ...)
}
\arguments{
\item{model}{A PVAR model}

\item{...}{Further arguments passed to or from other methods}
}
\description{
Sargan-Hansen-J-Test for Overidentification
}
\examples{
data("ex1_dahlberg_data")
hansen_j_test(ex1_dahlberg_data)
}
