% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvarfeols-methods.R
\name{knit_print.pvarfeols}
\alias{knit_print.pvarfeols}
\title{Knit Print Method for pvarfeols}
\usage{
knit_print.pvarfeols(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{further arguments}
}
\description{
Knit Print Method for pvarfeols
}
