% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvar-methods.R
\name{knit_print.pvargmm}
\alias{knit_print.pvargmm}
\title{Knit Print Method for pvargmm}
\usage{
knit_print.pvargmm(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{further arguments}
}
\description{
Knit Print Method for pvargmm
}
