% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvarhk-methods.R
\name{knit_print.pvarhk}
\alias{knit_print.pvarhk}
\title{Knit Print Method for pvarhk}
\usage{
knit_print.pvarhk(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{further arguments}
}
\description{
Knit Print Method for pvarhk
}
