% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvarfeols-methods.R
\name{knit_print.summary.pvarfeols}
\alias{knit_print.summary.pvarfeols}
\title{Knit Print summary Method}
\usage{
knit_print.summary.pvarfeols(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{further arguments}
}
\description{
Knit Print summary Method
}
