% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvar-methods.R
\name{knit_print.summary.pvargmm}
\alias{knit_print.summary.pvargmm}
\title{Knit Print summary Method}
\usage{
\method{knit_print}{summary.pvargmm}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{further arguments}
}
\description{
Knit Print summary Method
}
