% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafiles.R
\docType{data}
\name{nlswork2}
\alias{nlswork2}
\title{NLS Work 2 data}
\format{
An object of class \code{data.frame} with 16094 rows and 21 columns.
}
\usage{
nlswork2
}
\description{
NLS Work 2 data
}
\keyword{datasets}
