% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvar-methods.R
\name{oirf}
\alias{oirf}
\title{Orthogonal Impulse Response Function}
\usage{
oirf(model, n.ahead)
}
\arguments{
\item{model}{A PVAR model}

\item{n.ahead}{Any stable AR() model has an infinite MA representation. Hence any shock can be simulated infinitely into the future. For each forecast step t you need an addtional MA term.}
}
\description{
Orthogonal Impulse Response Function
}
