% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvar-methods.R
\name{plot.pvarstability}
\alias{plot.pvarstability}
\title{S3 plot method for pvarstability object, returns a \code{ggplot} object}
\usage{
\method{plot}{pvarstability}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{further arguments}
}
\description{
S3 plot method for pvarstability object, returns a \code{ggplot} object
}
