% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvarhk-methods.R
\name{print.pvarhk}
\alias{print.pvarhk}
\title{S3 Print Method for pvarhk}
\usage{
\method{print}{pvarhk}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{further arguments}
}
\description{
S3 Print Method for pvarhk
}
