% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvarfeols-methods.R
\name{print.summary.pvarfeols}
\alias{print.summary.pvarfeols}
\title{S3 Print Method for summary.pvarfeols}
\usage{
\method{print}{summary.pvarfeols}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{further arguments}
}
\description{
S3 Print Method for summary.pvarfeols
}
