% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvar-methods.R
\name{print.summary.pvargmm}
\alias{print.summary.pvargmm}
\title{S3 Print Method for summary.pvargmm}
\usage{
\method{print}{summary.pvargmm}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{further arguments}
}
\description{
S3 Print Method for summary.pvargmm
}
