% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvarhk-methods.R
\name{print.summary.pvarhk}
\alias{print.summary.pvarhk}
\title{S3 Print Method for summary.pvarhk}
\usage{
\method{print}{summary.pvarhk}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{further arguments}
}
\description{
S3 Print Method for summary.pvarhk
}
