% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvar-methods.R, R/pvarfeols-methods.R,
%   R/pvarhk-methods.R
\name{pvalue}
\alias{pvalue}
\alias{pvalue.pvargmm}
\alias{pvalue.pvarfeols}
\alias{pvalue.pvarhk}
\title{P-value S3 Method}
\usage{
pvalue(object, ...)

\method{pvalue}{pvargmm}(object, ...)

\method{pvalue}{pvarfeols}(object, ...)

\method{pvalue}{pvarhk}(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Further arguments}
}
\description{
P-value S3 Method
}
\examples{
data("ex1_dahlberg_data")
pvalue(ex1_dahlberg_data)
}
