% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvar-methods.R
\name{residuals_level}
\alias{residuals_level}
\alias{residuals_level.pvargmm}
\title{Extracting Level Residuals}
\usage{
residuals_level(model, ...)

\method{residuals_level}{pvargmm}(model, ...)
}
\arguments{
\item{model}{Model}

\item{...}{Further arguments passed to or from other methods}
}
\description{
Extracting Level Residuals
}
\examples{
data("ex1_dahlberg_data")
residuals_level(ex1_dahlberg_data)
}
