% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvar-methods.R, R/pvarfeols-methods.R
\name{stability}
\alias{stability}
\alias{stability.pvargmm}
\alias{stability.pvarfeols}
\title{Stability of PVAR(p) model}
\usage{
stability(model, ...)

\method{stability}{pvargmm}(model, ...)

\method{stability}{pvarfeols}(model, ...)
}
\arguments{
\item{model}{PVAR model}

\item{...}{Further arguments}
}
\value{
A \code{pvarstability} object containing eigenvalue stability conditions
}
\description{
Stability of PVAR(p) model
}
\examples{
data("ex1_dahlberg_data")
stability_info <- stability(ex1_dahlberg_data)
print(stability_info)
plot(stability_info)
}
