% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvarfeols-methods.R
\name{summary.pvarfeols}
\alias{summary.pvarfeols}
\title{S3 Summary Method for pvarfeols}
\usage{
\method{summary}{pvarfeols}(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{further arguments}
}
\description{
S3 Summary Method for pvarfeols
}
