% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvar-methods.R
\name{summary.pvargmm}
\alias{summary.pvargmm}
\title{S3 Summary Method for pvargmm}
\usage{
\method{summary}{pvargmm}(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{further arguments}
}
\description{
S3 Summary Method for pvargmm
}
