% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvarhk-methods.R
\name{summary.pvarhk}
\alias{summary.pvarhk}
\title{S3 Summary Method for pvarhk}
\usage{
\method{summary}{pvarhk}(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{further arguments}
}
\description{
S3 Summary Method for pvarhk
}
