% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvar-methods.R, R/pvarfeols-methods.R,
%   R/pvarhk-methods.R
\name{se}
\alias{se}
\alias{se.pvargmm}
\alias{se.pvarfeols}
\alias{se.pvarhk}
\title{Standard Error S3 Method}
\usage{
se(object, ...)

\method{se}{pvargmm}(object, ...)

\method{se}{pvarfeols}(object, ...)

\method{se}{pvarhk}(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Further arguments}
}
\description{
Standard Error S3 Method
}
\examples{
data("ex1_dahlberg_data")
se(ex1_dahlberg_data)
}
